requirements.txt
appdirs==1.4.4
argcomplete==3.6.2
asgiref==3.8.1
certifi==2025.4.26
charset-normalizer==3.4.2
colorama==0.4.6
Django==5.2.1
django-admin-rangefilter==0.13.3
django-cors-headers==4.7.0
django-filter==25.1
django-stubs==5.2.5
django-stubs-ext==5.2.5
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
dotenv==0.9.9
drf-yasg==1.21.10
idna==3.10
importlib_resources==6.5.2
inflection==0.5.1
mypy_extensions==1.1.0
numpy==2.3.3
packaging==25.0
pandas==2.3.2
pillow==11.2.1
psycopg2-binary==2.9.10
PyJWT==2.9.0
python-barcode==0.15.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-escpos==3.1
pytz==2025.2
PyYAML==6.0.2
qrcode==8.2
reportlab==4.4.3
requests==2.32.3
setuptools==80.8.0
six==1.17.0
sqlparse==0.5.3
types-PyYAML==6.0.12.20250915
typing_extensions==4.15.0
tzdata==2025.2
uritemplate==4.2.0
urllib3==2.4.0
xlsxwriter==3.2.9
